+++ src/ldc/supervised/pairs/__init__.py
@@ -4,4 +4,5 @@ from ._alpaca import AlpacaReader, AlpacaWriter
 from ._csv import CsvPairsReader, CsvPairsWriter, TsvPairsReader, TsvPairsWriter
 from ._jsonlines import JsonLinesPairReader, JsonLinesPairWriter
 from ._parquet import ParquetPairsReader, ParquetPairsWriter
+from ._update_pair_data import UpdatePairData
 from ._xtuner import XTunerReader, XTunerWriter